+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-01-04
+
+### Added
+
+#### String Utilities
+- **padZero**: Pads a number with leading 0s to match a certain length
+
+#### Number Utilities
+- **clamp**: Bounds a given number to a defined range
+- **modulo**: Performs a modulo (%) operation whose result is guaranteed to be positive
+- **randomInteger**: Generates a random integer between 0 and a specified maximum (exclusive)
+
+#### Array Utilities
+- **isEquivalent**: Checks whether two different arrays have equivalent contents
+- **removeAllInPlace**: Removes all occurrences of an element from an array in-place
+- **removeInPlace**: Removes the first occurrence of an element from an array in-place
+
 ## [1.0.0] - 2025-11-29
 
 ### Added